Broken bidirectional.
While it's possible to fetch all the lineage of a node in a single SQL query
the prefetching framework does not allow for multiple relationships (forward
and reverse in our case) to be applied to top level objects.

In order to fetch that object assignment and other internal parts of the
prefetching machinery should delegate to Prefetch methods.