Deprecated DEFAULT_AUTO_FIELD in favor of DEFAULT_PK_FIELD.
Now that we fully support Field.db_default there is no reasons to restrict
default primary keys to ones that subclass AutoField as any field able to
manage its own value generation is suitable as an implicit primary key.

The usage of *auto* in the feature name was ambiguous as it referred both to
AutoField which gets its name from the MySQL AUTO_INCREMENT implementation of
serial fields and the automatic definition of id fields on models.